requirements.txt
loguru==0.7.2
paho-mqtt==1.6.1
pydantic==2.4.2
pymodbus==3.5.4
PyYAML==6.0.1
